Boss Implementing into main scene 2/2
Things still to be done :

- Player Knockback
- Player Hitting the minotaur
- Quest Colliders

+++ Project T/Assets/_Scripts/_FrontEnd/_UI/HUDManager.cs
@@ -39,6 +39,7 @@ public class HUDManager : MonoBehaviour {
 
         itObject = GameObject.FindGameObjectWithTag("itObject");
 
+       
         playerHPCanvas = playerHP.GetComponent<CanvasGroup>();
         bossHPCanvas = bossHP.GetComponent<CanvasGroup>();
         bossHPCanvas.alpha = 0;
@@ -72,7 +73,7 @@ public class HUDManager : MonoBehaviour {
 
 
     }
-    public void UpdateBossHealth(int health) {
+    public void UpdateBossHealth(float health) {
         if (bossHPCanvas.alpha == 0) {
             bossHPCanvas.alpha = 1;
         }
